+++ src/routes/all.ts
@@ -10,7 +10,7 @@ const gymapp_api = apiAdapter(BASE_URL);
 allRouter.get('/desafios/:desafioId/submissoes', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
   const desafioId = req.params.desafioId;
-  gymapp_api.get(`/desafios/${desafioId}/submissoes`,
+  gymapp_api.get(`/${userId}/desafios/${desafioId}/submissoes`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -19,7 +19,7 @@ allRouter.get('/desafios/:desafioId/submissoes', verificarAutenticacao, (req, re
 allRouter.get('/alunos/ginasio/:ginasioId', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
   const ginasioId = req.params.ginasioId;
-  gymapp_api.get(`/alunos/ginasio/${ginasioId}`,
+  gymapp_api.get(`/${userId}/alunos/ginasio/${ginasioId}`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -30,7 +30,7 @@ allRouter.post('/posts/comentarios', verificarAutenticacao, (req, res) => {
   const postId = req.body.postId;
   let body = req.body;
   delete body['postId'];
-  gymapp_api.post(`/posts/${postId}/comentarios`,
+  gymapp_api.post(`/${userId}/posts/${postId}/comentarios`,
     body
   ).then(resp => {
     res.send(resp.data)
@@ -41,7 +41,7 @@ allRouter.post('/posts/comentario/gosto', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
   const postId = req.body.postId;
   const comentarioId = req.body.comentarioId;
-  gymapp_api.post(`/posts/${postId}/comentario/${comentarioId}/gosto`,
+  gymapp_api.post(`/${userId}/posts/${postId}/comentario/${comentarioId}/gosto`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -51,7 +51,7 @@ allRouter.delete('/posts/:postId/comentario/:comentarioId/gosto', verificarAuten
   const userId = res.locals.uid;
   const postId = req.params.postId;
   const comentarioId = req.params.comentarioId;
-  gymapp_api.delete(`/posts/${postId}/comentario/${comentarioId}/gosto`,
+  gymapp_api.delete(`/${userId}/posts/${postId}/comentario/${comentarioId}/gosto`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -62,7 +62,7 @@ allRouter.delete('/posts/:postId/comentario/:comentarioId', verificarAutenticaca
   const postId = req.params.postId;
   const comentarioId = req.params.comentarioId;
 
-  gymapp_api.delete(`/posts/${postId}/comentario/${comentarioId}`,
+  gymapp_api.delete(`/${userId}/posts/${postId}/comentario/${comentarioId}`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -71,7 +71,7 @@ allRouter.delete('/posts/:postId/comentario/:comentarioId', verificarAutenticaca
 allRouter.get('/posts/:postId', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
   const postId = req.params.postId;
-  gymapp_api.get(`/posts/${postId}`,
+  gymapp_api.get(`/${userId}/posts/${postId}`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -80,7 +80,7 @@ allRouter.get('/posts/:postId', verificarAutenticacao, (req, res) => {
 allRouter.delete('/posts/:postId', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
   const postId = req.params.postId;
-  gymapp_api.delete(`/posts/${postId}`,
+  gymapp_api.delete(`/${userId}/posts/${postId}`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -91,7 +91,7 @@ allRouter.put('/posts', verificarAutenticacao, (req, res) => {
   const postId = req.body.postId;
   let body = req.body;
   delete body['postId'];
-  gymapp_api.put(`/posts/${postId}`,
+  gymapp_api.put(`/${userId}/posts/${postId}`,
     body
   ).then(resp => {
     res.send(resp.data)
@@ -103,7 +103,7 @@ allRouter.post('/posts/gostos', verificarAutenticacao, (req, res) => {
   const postId = req.body.postId;
   let body = req.body;
   delete body['postId'];
-  gymapp_api.post(`/posts/${postId}/gostos`,
+  gymapp_api.post(`/${userId}/posts/${postId}/gostos`,
     body
   ).then(resp => {
     res.send(resp.data)
@@ -113,7 +113,7 @@ allRouter.post('/posts/gostos', verificarAutenticacao, (req, res) => {
 allRouter.delete('/posts/:postId/gostos', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
   const postId = req.params.postId;
-  gymapp_api.delete(`/posts/${postId}/gostos`,
+  gymapp_api.delete(`/${userId}/posts/${postId}/gostos`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -122,7 +122,7 @@ allRouter.delete('/posts/:postId/gostos', verificarAutenticacao, (req, res) => {
 allRouter.get('/desafios/:desafioId', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
   const desafioId = req.params.desafioId;
-  gymapp_api.get(`/desafios/${desafioId}`,
+  gymapp_api.get(`/${userId}/desafios/${desafioId}`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -130,7 +130,7 @@ allRouter.get('/desafios/:desafioId', verificarAutenticacao, (req, res) => {
 
 allRouter.get('/atividades', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
-  gymapp_api.get(`/atividades/`,
+  gymapp_api.get(`/${userId}/atividades/`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -138,7 +138,7 @@ allRouter.get('/atividades', verificarAutenticacao, (req, res) => {
 
 allRouter.get('/definicoes', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
-  gymapp_api.get(`/definicoes`,
+  gymapp_api.get(`/${userId}/definicoes`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -147,7 +147,7 @@ allRouter.get('/definicoes', verificarAutenticacao, (req, res) => {
 allRouter.put('/destinosNotificacao/notificacao', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
   const notiId = req.body.notiId;
-  gymapp_api.put(`/destinosNotificacao/notificacao/${notiId}`,
+  gymapp_api.put(`/${userId}/destinosNotificacao/notificacao/${notiId}`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -155,7 +155,7 @@ allRouter.put('/destinosNotificacao/notificacao', verificarAutenticacao, (req, r
 
 allRouter.put('/perfil', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
-  gymapp_api.put(`/perfil`,
+  gymapp_api.put(`/${userId}/perfil`,
     req.body
   ).then(resp => {
     res.send(resp.data)
@@ -164,7 +164,7 @@ allRouter.put('/perfil', verificarAutenticacao, (req, res) => {
 
 allRouter.get('/perfil', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
-  gymapp_api.get(`/perfil`,
+  gymapp_api.get(`/${userId}/perfil`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -173,7 +173,7 @@ allRouter.get('/perfil', verificarAutenticacao, (req, res) => {
 allRouter.get('/user/:uId', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
   const uId = req.params.uId;
-  gymapp_api.get(`/user/${uId}`,
+  gymapp_api.get(`/${userId}/user/${uId}`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -181,7 +181,7 @@ allRouter.get('/user/:uId', verificarAutenticacao, (req, res) => {
 
 allRouter.put('/definicoes/identificacao', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
-  gymapp_api.put(`/definicoes/identificacao/`,
+  gymapp_api.put(`/${userId}/definicoes/identificacao/`,
     req.body
   ).then(resp => {
     res.send(resp.data)
@@ -190,7 +190,7 @@ allRouter.put('/definicoes/identificacao', verificarAutenticacao, (req, res) => 
 
 allRouter.put('/definicoes/mencoes', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
-  gymapp_api.put(`/definicoes/mencoes`,
+  gymapp_api.put(`/${userId}/definicoes/mencoes`,
     req.body
   ).then(resp => {
     res.send(resp.data)
@@ -199,7 +199,7 @@ allRouter.put('/definicoes/mencoes', verificarAutenticacao, (req, res) => {
 
 allRouter.get('/posts', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
-  gymapp_api.get(`/posts`,
+  gymapp_api.get(`/${userId}/posts`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -207,7 +207,7 @@ allRouter.get('/posts', verificarAutenticacao, (req, res) => {
 
 allRouter.post('/posts', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
-  gymapp_api.post(`/posts`,
+  gymapp_api.post(`/${userId}/posts`,
     req.body
   ).then(resp => {
     res.send(resp.data)
@@ -216,7 +216,7 @@ allRouter.post('/posts', verificarAutenticacao, (req, res) => {
 
 allRouter.put('/definicoes/perfil/privado', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
-  gymapp_api.put(`/definicoes/perfil/privado/`,
+  gymapp_api.put(`/${userId}/definicoes/perfil/privado/`,
     req.body
   ).then(resp => {
     res.send(resp.data)
@@ -226,7 +226,7 @@ allRouter.put('/definicoes/perfil/privado', verificarAutenticacao, (req, res) =>
 allRouter.get('/ginasio/:ginasioId/desafios', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
   const ginasioId = req.params.ginasioId;
-  gymapp_api.get(`/ginasio/${ginasioId}/desafios/`,
+  gymapp_api.get(`/${userId}/ginasio/${ginasioId}/desafios/`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -235,7 +235,7 @@ allRouter.get('/ginasio/:ginasioId/desafios', verificarAutenticacao, (req, res) 
 allRouter.get('/ginasio/:ginasioId/desafios/disponiveis', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
   const ginasioId = req.params.ginasioId;
-  gymapp_api.get(`/ginasio/${ginasioId}/desafios/disponiveis`,
+  gymapp_api.get(`/${userId}/ginasio/${ginasioId}/desafios/disponiveis`,
   ).then(resp => {
     res.send(resp.data)
   })
@@ -243,7 +243,7 @@ allRouter.get('/ginasio/:ginasioId/desafios/disponiveis', verificarAutenticacao,
 
 allRouter.get('/notificacoes', verificarAutenticacao, (req, res) => {
   const userId = res.locals.uid;
-  gymapp_api.get(`/notificacoes`,
+  gymapp_api.get(`/${userId}/notificacoes`,
   ).then(resp => {
     res.send(resp.data)
   })
